requirements.txt
lxml
requests
unicodecsv
argparse
bs4
datetime
pandas
numpy
websocket
websocket-client
asyncio
